requirements.txt
# Dependencias para ETL con Corda
requests==2.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
json
datetime
logging
configparser
corda-client==1.0.0
websocket-client==1.6.4

# Dependencias para Dashboard con IA
streamlit==1.28.1
plotly==5.17.0
scikit-learn==1.3.2
prophet==1.1.4
pycaret==3.0.4
numpy==1.24.3
seaborn==0.12.2
matplotlib==3.7.2
